Move getArchivePosts to functions
