+++ pages/team/[[...slug]].js
@@ -6,9 +6,9 @@ import Layout from '@/components/common/Layout'
 import Blocks from '@/components/molecules/Blocks'
 import Card from '@/components/molecules/Card'
 import getPagePropTypes from '@/functions/getPagePropTypes'
+import getArchivePosts from '@/functions/next-api/wordpress/archive/getArchivePosts'
 import getPostTypeStaticPaths from '@/functions/wordpress/postTypes/getPostTypeStaticPaths'
 import getPostTypeStaticProps from '@/functions/wordpress/postTypes/getPostTypeStaticProps'
-import getArchivePosts from '@/lib/next-api/wordpress/archive/getArchivePosts'
 
 // Define route post type.
 const postType = 'team'
